requirements.txt
kafka-python==2.0.2
boto3==1.16.26
pandas==1.1.4
certifi==2020.11.8
retry==0.9.2
mysql-connector-python==8.0.22